requirements.txt
tensorflow
numpy
Pillow
gradio
scipy